pyproject.toml
[project]
name = "discord-search"
version = "0.1.0"
description = "Discord会話を自然言語で検索できるBot"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.3.0",
    "google-genai>=1.0.0",
    "google-cloud-firestore>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
ocr = [
    "yomitoku>=0.10.0",
    "torch>=2.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
